Move umap-specific stuff into its own file
